Refactor navigation, use parcelable for trail data, and implement back button on details screen.
