+++ app/src/main/java/com/example/gps/MainActivity.kt
@@ -1,7 +1,6 @@
 package com.example.gps
 
 import android.Manifest
-import android.content.Intent
 import android.content.pm.PackageManager
 import android.os.Bundle
 import android.widget.Toast
@@ -13,12 +12,9 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.core.content.ContextCompat
-import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import androidx.navigation.NavType
-import androidx.navigation.navArgument
 import androidx.navigation.compose.rememberNavController
 import com.google.firebase.FirebaseApp
 import com.google.firebase.auth.FirebaseAuth
@@ -26,10 +22,9 @@ import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
 import com.example.gps.navigation.BottomNavBar
 import android.net.Uri
-import com.example.gps.MapScreen
-import com.example.gps.TrailCreationScreen
 import com.google.firebase.appcheck.FirebaseAppCheck
 import com.google.firebase.appcheck.playintegrity.PlayIntegrityAppCheckProviderFactory
+import com.example.gps.data.Trail
 
 class MainActivity : ComponentActivity() {
 
@@ -113,27 +108,16 @@ fun NavHostContainer(navController: NavHostController, modifier: Modifier) {
         composable("create") { TrailCreationScreen( navController) }
         composable("browse") {
             BrowseScreen { trail ->
-                navController.navigate(
-                    "details/${Uri.encode(trail.name)}/${Uri.encode(trail.location)}/${Uri.encode(trail.difficulty)}/${Uri.encode(trail.imageUrl)}"
-                )
+                navController.currentBackStackEntry?.savedStateHandle?.set("selectedTrail", trail)
+                navController.navigate("details")
             }
         }
-        composable(
-            "details/{name}/{location}/{difficulty}/{imageUrl}",
-            arguments = listOf(
-                navArgument("name") { type = NavType.StringType },
-                navArgument("location") { type = NavType.StringType },
-                navArgument("difficulty") { type = NavType.StringType },
-                navArgument("imageUrl") { type = NavType.StringType }
-            )
-        ) { backStackEntry ->
-            val name = Uri.decode(backStackEntry.arguments?.getString("name") ?: "")
-            val location = Uri.decode(backStackEntry.arguments?.getString("location") ?: "")
-            val difficulty = Uri.decode(backStackEntry.arguments?.getString("difficulty") ?: "")
-            val imageUrl = backStackEntry.arguments?.getString("imageUrl") ?: ""
+        composable("details") { backStackEntry ->
+            val trail = navController.previousBackStackEntry
+                ?.savedStateHandle
+                ?.get<Trail>("selectedTrail")
 
-            val trail = Trail(name, location, difficulty, imageUrl)
-            TrailDetailsScreen(trail)
+            trail?.let { TrailDetailsScreen(it, navController) }
         }
         composable("waypointSelection") { WaypointSelectionScreen(navController) }
     }
