+++ app/src/main/java/com/example/gps/TrailDetailsScreen.kt
@@ -3,6 +3,8 @@ package com.example.gps
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -11,11 +13,18 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
+import androidx.navigation.NavController
 import coil3.compose.rememberAsyncImagePainter
+import com.example.gps.data.Trail
 
 @Composable
-fun TrailDetailsScreen(trail: Trail) {
+fun TrailDetailsScreen(trail: Trail,navController: NavController) {
+
+    IconButton(onClick = { navController.popBackStack()  }) {
+        Icon(Icons.AutoMirrored.Filled.ArrowBack, contentDescription = "Back")
+    }
+    Spacer(modifier = Modifier.height(16.dp))
+
     Column(
         modifier = Modifier
             .fillMaxSize()
